Breaking setup and goals into abstract/concrete classes.
